feat: 서버 재시작 시 오브젝트 데이터로 메모리에 로드
